+++ common/src/main/kotlin/com/cobblemon/mod/common/command/TestCommand.kt
@@ -8,28 +8,17 @@
 
 package com.cobblemon.mod.common.command
 
-import com.cobblemon.mod.common.CobblemonItems
-import com.cobblemon.mod.common.api.pokemon.PokemonSpecies
-import com.cobblemon.mod.common.battles.BattleFormat
-import com.cobblemon.mod.common.battles.BattleRegistry
-import com.cobblemon.mod.common.battles.BattleSide
-import com.cobblemon.mod.common.battles.actor.MultiPokemonBattleActor
-import com.cobblemon.mod.common.battles.actor.PlayerBattleActor
-import com.cobblemon.mod.common.battles.pokemon.BattlePokemon
-import com.cobblemon.mod.common.item.PokemonItem
-import com.cobblemon.mod.common.pokemon.Pokemon
-import com.cobblemon.mod.common.util.cobblemonResource
-import com.cobblemon.mod.common.util.party
+import com.cobblemon.mod.common.battles.runner.GraalShowdown
+import com.google.gson.GsonBuilder
+import com.google.gson.JsonObject
 import com.mojang.brigadier.Command
 import com.mojang.brigadier.CommandDispatcher
 import com.mojang.brigadier.context.CommandContext
-import net.minecraft.item.ItemStack
-import net.minecraft.nbt.NbtList
-import net.minecraft.nbt.NbtString
+import java.io.File
+import java.io.PrintWriter
 import net.minecraft.server.command.CommandManager
 import net.minecraft.server.command.ServerCommandSource
 import net.minecraft.server.network.ServerPlayerEntity
-import net.minecraft.server.world.ServerWorld
 
 object TestCommand {
 
@@ -46,6 +35,7 @@ object TestCommand {
         }
 
         try {
+//            extractMovesData()
 //            // Player variables
 //            val player = context.source.entity as ServerPlayerEntity
 //            val party = player.party()
@@ -76,11 +66,71 @@ object TestCommand {
 //                side2 = BattleSide(MultiPokemonBattleActor(listOf(enemyPokemon, enemyPokemon2, enemyPokemon3, enemyPokemon4)))
 //            )
 
-            val player = context.source.entity as ServerPlayerEntity
-            player.giveItemStack(PokemonItem.from(PokemonSpecies.random(), "alolan"))
+//            val player = context.source.entity as ServerPlayerEntity
+//            player.giveItemStack(PokemonItem.from(PokemonSpecies.random(), "alolan"))
         } catch (e: Exception) {
             e.printStackTrace()
         }
         return Command.SINGLE_SUCCESS
+    }
+
+    private fun extractMovesData() {
+        val ctx = GraalShowdown.context
+        ctx.eval("js", """
+                const ShowdownMoves = require('pokemon-showdown/data/moves');
+            """.trimIndent())
+        val moves = ctx.getBindings("js").getMember("ShowdownMoves").getMember("Moves")
+        val gson = GsonBuilder().setPrettyPrinting().create()
+        for (moveName in moves.memberKeys) {
+            try {
+                val value = moves.getMember(moveName)
+                val obj = JsonObject()
+                obj.addProperty("name", moveName)
+                obj.addProperty("type", value.getMember("type").asString())
+                obj.addProperty("damageCategory", value.getMember("category").asString())
+                obj.addProperty("target", value.getMember("target").asString())
+                obj.addProperty("power", value.getMember("basePower").asInt())
+                obj.addProperty("accuracy", value.getMember("accuracy").let { if (it.isBoolean) -1F else it.asFloat() })
+                obj.addProperty("pp", value.getMember("pp").asInt())
+                obj.addProperty("priority", value.getMember("priority").asInt())
+                if (value.hasMember("secondary")) {
+                    val secondary = value.getMember("secondary")
+                    if (secondary.hasMember("chance")) {
+                        obj.addProperty("effectChance", secondary.getMember("chance").asInt())
+                    }
+                }
+                val file = File("outputmoves").also { it.mkdir() }
+                val pw = PrintWriter(File(file, "$moveName.json"))
+                pw.write(gson.toJson(obj))
+                pw.close()
+            } catch (e: Exception) {
+                println("Issue when converting $moveName")
+                e.printStackTrace()
+            }
+        }
+    }
+
+    private fun extractAbilitiesData() {
+        val ctx = GraalShowdown.context
+        ctx.eval("js", """
+            const ShowdownAbilities = require('pokemon-showdown/data/abilities');
+        """.trimIndent())
+        val abilities = ctx.getBindings("js").getMember("ShowdownAbilities").getMember("Abilities")
+        val gson = GsonBuilder().setPrettyPrinting().create()
+        for (abilityName in abilities.memberKeys) {
+            try {
+                val obj = JsonObject()
+                obj.addProperty("name", abilityName)
+                obj.addProperty("displayName", "cobblemon.ability.$abilityName")
+                obj.addProperty("description", "cobblemon.ability.$abilityName.desc")
+                val file = File("outputabilities").also { it.mkdir() }
+                val pw = PrintWriter(File(file, "$abilityName.json"))
+                pw.write(gson.toJson(obj))
+                pw.close()
+            } catch (e: Exception) {
+                println("Issue when converting $abilityName")
+                e.printStackTrace()
+            }
+        }
     }
 }